# openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SmartHotel OS Enterprise API Catalog",
    "description": "Standardized REST API specifications, query parameters, payload models, and RBAC requirements for the SmartHotel OS hospitality platform.",
    "version": "1.0.0",
    "contact": {
      "name": "SmartHotel Operations Engineering Team",
      "email": "engineering@smarthotel.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local / Production relative root path"
    }
  ],
  "paths": {
...
    "/complaints": {
      "get": {
        "summary": "Retrieve Complaints",
        "description": "Returns a list of complaints. Non-admins only receive their own complaints. Admins can filter by userId.",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by userId (admin only)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of complaints",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Complaint"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a Complaint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplaintRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Complaint created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Complaint"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZodValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Cannot link complaint to an unowned booking"
          }
        },
        "description": "Creates a new complaint. If bookingId is provided, it must belong to the authenticated user."
      },
      "patch": {
        "summary": "Update a Complaint",
        "description": "Admin only endpoint to update complaint status.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateComplaintRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Complaint updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Complaint"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZodValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Must be an admin"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "NextAuthCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      },
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ZodValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation error"
          },
          "details": {
            "type": "object",
            "description": "Zod-formatted error details object."
          }
        }
      },
...
      "Complaint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subject": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED",
              "CLOSED"
            ]
          },
          "bookingId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CreateComplaintRequest": {
        "type": "object",
        "required": [
          "subject",
          "description",
          "category"
        ],
        "properties": {
          "subject": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "minLength": 10
          },
          "category": {
            "type": "string",
            "minLength": 2
          },
          "bookingId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          }
        }
      },
      "UpdateComplaintRequest": {
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED",
              "CLOSED"
            ]
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  }
}